Revert "Release v2.0.0 (#545)"
This reverts commit df6a17f1d9ed8c0d37c6b6cc1daf90e42884afe0.
