+++ configs/logger/csv.yaml
@@ -1,7 +1,7 @@
 # csv logger built in lightning
 
 csv:
-  _target_: lightning.pytorch.loggers.csv_logs.CSVLogger
+  _target_: pytorch_lightning.loggers.csv_logs.CSVLogger
   save_dir: "${paths.output_dir}"
   name: "csv/"
   prefix: ""
