+++ configs/logger/mlflow.yaml
@@ -1,7 +1,7 @@
 # https://mlflow.org
 
 mlflow:
-  _target_: lightning.pytorch.loggers.mlflow.MLFlowLogger
+  _target_: pytorch_lightning.loggers.mlflow.MLFlowLogger
   # experiment_name: ""
   # run_name: ""
   tracking_uri: ${paths.log_dir}/mlflow/mlruns # run `mlflow ui` command inside the `logs/mlflow/` dir to open the UI
