+++ configs/logger/neptune.yaml
@@ -1,7 +1,7 @@
 # https://neptune.ai
 
 neptune:
-  _target_: lightning.pytorch.loggers.neptune.NeptuneLogger
+  _target_: pytorch_lightning.loggers.neptune.NeptuneLogger
   api_key: ${oc.env:NEPTUNE_API_TOKEN} # api key is loaded from environment variable
   project: username/lightning-hydra-template
   # name: ""
