+++ configs/logger/wandb.yaml
@@ -1,7 +1,7 @@
 # https://wandb.ai
 
 wandb:
-  _target_: lightning.pytorch.loggers.wandb.WandbLogger
+  _target_: pytorch_lightning.loggers.wandb.WandbLogger
   # name: "" # name of the run (normally generated by wandb)
   save_dir: "${paths.output_dir}"
   offline: False
