+++ configs/trainer/default.yaml
@@ -1,4 +1,4 @@
-_target_: lightning.pytorch.trainer.Trainer
+_target_: pytorch_lightning.Trainer
 
 default_root_dir: ${paths.output_dir}
