+++ src/utils/__init__.py
@@ -1,5 +1,12 @@
-from src.utils.instantiatiators import instantiate_callbacks, instantiate_loggers
-from src.utils.logging_utils import log_hyperparameters
 from src.utils.pylogger import get_pylogger
 from src.utils.rich_utils import enforce_tags, print_config_tree
-from src.utils.utils import extras, get_metric_value, task_wrapper
+from src.utils.utils import (
+    close_loggers,
+    extras,
+    get_metric_value,
+    instantiate_callbacks,
+    instantiate_loggers,
+    log_hyperparameters,
+    save_file,
+    task_wrapper,
+)
